Move cleanup logic for old rcache data to rcache package
