+++ pkg/rcache/rcache.go
@@ -18,6 +18,13 @@ import (
 // effectively resetting the cache at the same time the new code
 // is deployed.
 const dataVersion = "v2"
+const dataVersionToDelete = "v1"
+
+// DeleteOldCacheData deletes the rcache data in the given Redis instance
+// that's prefixed with dataVersionToDelete
+func DeleteOldCacheData(c redis.Conn) error {
+	return deleteKeysWithPrefix(c, dataVersionToDelete)
+}
 
 // Cache implements httpcache.Cache
 type Cache struct {
@@ -205,6 +212,62 @@ end`, 0, globalPrefix+":*")
 	if err != nil {
 		log15.Error("Could not clear test prefix", "name", t.Name(), "globalPrefix", globalPrefix, "error", err)
 	}
+}
+
+func deleteKeysWithPrefix(c redis.Conn, prefix string) error {
+	pattern := prefix + ":*"
+
+	iter := 0
+	keys := make([]string, 0)
+	for {
+		arr, err := redis.Values(c.Do("SCAN", iter, "MATCH", pattern))
+		if err != nil {
+			return fmt.Errorf("error retrieving keys with pattern %q", pattern)
+		}
+
+		iter, err = redis.Int(arr[0], nil)
+		if err != nil {
+			return err
+		}
+
+		k, err := redis.Strings(arr[1], nil)
+		if err != nil {
+			return err
+		}
+		keys = append(keys, k...)
+		if iter == 0 {
+			break
+		}
+	}
+
+	if len(keys) == 0 {
+		return nil
+	}
+
+	const batchSize = 1000
+	var batch = make([]interface{}, batchSize, batchSize)
+
+	for i := 0; i < len(keys); i += batchSize {
+		j := i + batchSize
+		if j > len(keys) {
+			j = len(keys)
+		}
+		currentBatchSize := j - i
+
+		for bi, v := range keys[i:j] {
+			batch[bi] = v
+		}
+
+		// We ignore whether the number of deleted keys matches what we have in
+		// `batch`, because in the time since we constructed `keys` some of the
+		// keys might have expired
+		_, err := c.Do("DEL", batch[:currentBatchSize]...)
+		if err != nil {
+			return fmt.Errorf("failed to delete keys: %s", err)
+		}
+	}
+
+	return nil
 }
 
 var (
